requirements.txt
apsw-wheels==3.26.0.post2
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
bottleneck==1.2.1
certifi==2019.6.16
chardet==3.0.4
click==7.0
cloudpickle==1.2.1
colorama==0.4.1
coverage==4.5.3
dash==1.0.2
-e git+https://github.com/KyleKing/dash_charts.git@4eaeb91ed895f61886e5c0a924db508062163c10#egg=dash-charts
dash-core-components==1.0.0
dash-html-components==1.0.0
dash-renderer==1.0.0
dash-table==4.0.2
doit==0.31.1
entrypoints==0.3
flake8==3.7.8
flake8-blind-except==0.1.1
flake8-bugbear==19.3.0
flake8-builtins==1.4.1
flake8-commas==2.0.0
flake8-deprecated==1.3
flake8-isort==2.7.0
flake8-logging-format==0.6.0
flake8-mock==0.3
flake8-polyfill==1.0.2
flake8-print==3.1.0
flake8-quotes==2.0.1
flake8-tuple==0.4.0
flask==1.1.1
flask-compress==1.4.0
future==0.17.1
gitchangelog==3.0.4
icecream==1.5.0
idna==2.8
importlib-metadata==0.19
isort==4.3.21
itsdangerous==1.1.0
jinja2==2.10.1
macfsevents==0.8.1
mako==1.0.14
mando==0.6.4
markdown==3.1.1
markupsafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.0
packaging==19.0
pandas==0.24.2
pdoc3==0.6.3
pep8-naming==0.8.2
plaid-python==3.3.0
plaidash==1.0.0
plotly==4.0.0
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pygments==2.4.2
pyinotify==0.9.6
pyparsing==2.4.1.1
pytest==4.6.4
pytest-cov==2.7.1
python-dateutil==2.8.0
pytz==2019.1
pyyaml==5.1.1
radon==3.0.3
requests==2.22.0
retrying==1.3.3
six==1.12.0
testfixtures==6.10.0
toml==0.10.0
tqdm==4.32.2
untokenize==0.1.1
urllib3==1.25.3
wcwidth==0.1.7
werkzeug==0.15.5
zipp==0.5.2
